+++ src/Ext/Script/Body.cpp
@@ -221,6 +221,9 @@ void ScriptExt::ProcessAction(TeamClass* pTeam)
 	case PhobosScripts::SetMinimumAmmoThreshold:
 		ScriptExt::SetMinimumAmmoThreshold(pTeam, -2);
 		break;
+	case PhobosScripts::ForceGlobalOnlyTargetHouseEnemy:
+		ScriptExt::ForceGlobalOnlyTargetHouseEnemy(pTeam, -1);
+		break;
 	default:
 		// Do nothing because or it is a wrong Action number or it is an Ares/YR action...
 		if (action > 70 && !IsExtVariableAction(action))
@@ -1417,6 +1420,31 @@ void ScriptExt::SetMinimumAmmoThreshold(TeamClass* pTeam, int newValue = -2)
 		newValue = pTeam->CurrentScript->Type->ScriptActions[pTeam->CurrentScript->CurrentMission].Argument;
 
 	pTeamData->MinAmmoThreshold = newValue < 0 ? 0 : newValue;
+
+	// This action finished
+	pTeam->StepCompleted = true;
+}
+
+void ScriptExt::ForceGlobalOnlyTargetHouseEnemy(TeamClass* pTeam, int mode)
+{
+	if (!pTeam)
+		return;
+
+	auto pHouseExt = HouseExt::ExtMap.Find(pTeam->Owner);
+	if (!pHouseExt)
+	{
+		// This action finished
+		pTeam->StepCompleted = true;
+		return;
+	}
+
+	if (mode < 0 || mode > 2)
+		mode = pTeam->CurrentScript->Type->ScriptActions[pTeam->CurrentScript->CurrentMission].Argument;
+
+	if (mode < -1 || mode > 2)
+		mode = -1;
+
+	HouseExt::ForceOnlyTargetHouseEnemy(pTeam->Owner, mode);
 
 	// This action finished
 	pTeam->StepCompleted = true;